+++ python/cudf/cudf/core/udf/utils.py
@@ -8,10 +8,12 @@ import cachetools
 import cupy as cp
 import llvmlite.binding as ll
 import numpy as np
+from cubinlinker.patch import _numba_version_ok, get_logger, new_patched_linker
 from numba import cuda, typeof
 from numba.core.datamodel import default_manager
 from numba.core.errors import TypingError
 from numba.cuda.cudadrv import nvvm
+from numba.cuda.cudadrv.driver import Linker
 from numba.np import numpy_support
 from numba.types import CPointer, Poison, Tuple, boolean, int64, void
 
@@ -28,6 +30,9 @@ from cudf.utils.dtypes import (
     TIMEDELTA_TYPES,
 )
 from cudf.utils.utils import _cudf_nvtx_annotate
+
+logger = get_logger()
+
 
 JIT_SUPPORTED_TYPES = (
     NUMERIC_TYPES | BOOL_TYPES | DATETIME_TYPES | TIMEDELTA_TYPES
@@ -336,3 +341,83 @@ def _get_extensionty_size(ty):
     target_data = ll.create_target_data(data_layout)
     llty = default_manager[ty].get_value_type()
     return llty.get_abi_size(target_data)
+
+
+def _get_cuda_version_from_ptx_file(path):
+    """
+    https://docs.nvidia.com/cuda/parallel-thread-execution/
+    Each PTX module must begin with a .version
+    directive specifying the PTX language version
+
+    example header:
+    //
+    // Generated by NVIDIA NVVM Compiler
+    //
+    // Compiler Build ID: CL-31057947
+    // Cuda compilation tools, release 11.6, V11.6.124
+    // Based on NVVM 7.0.1
+    //
+
+    .version 7.6
+    .target sm_52
+    .address_size 64
+
+    """
+    with open(path) as ptx_file:
+        for line in ptx_file:
+            if line.startswith(".version"):
+                ver_line = line
+                break
+        else:
+            raise ValueError("Could not read CUDA version from ptx file.")
+    version = ver_line.strip("\n").split(" ")[1]
+    # from ptx_docs/release_notes above:
+    ver_map = {
+        "7.5": (11, 5),
+        "7.6": (11, 6),
+        "7.7": (11, 7),
+        "7.8": (11, 8),
+        "8.0": (12, 0),
+    }
+
+    cuda_ver = ver_map.get(version)
+    if cuda_ver is None:
+        raise ValueError(
+            f"Could not map PTX version {version} to a CUDA version"
+        )
+
+    return cuda_ver
+
+
+def _setup_numba_linker(path, prefix):
+    from ptxcompiler.patch import NO_DRIVER, safe_get_versions
+
+    from cudf.core.udf.utils import (
+        _get_cuda_version_from_ptx_file,
+        _get_ptx_file,
+        maybe_patch_numba_linker,
+    )
+
+    versions = safe_get_versions()
+    if versions != NO_DRIVER:
+        driver_version, runtime_version = versions
+        ptxpath = _get_ptx_file(path, prefix)
+        strings_udf_ptx_version = _get_cuda_version_from_ptx_file(ptxpath)
+        maybe_patch_numba_linker(driver_version, strings_udf_ptx_version)
+
+
+def maybe_patch_numba_linker(driver_version, ptx_toolkit_version):
+    # Numba thinks cubinlinker is only needed if the driver is older than
+    # the ctk but when strings_udf is present, it might also need to patch
+    # because the PTX file strings_udf relies on may be newer than
+    # the driver as well
+    if driver_version < ptx_toolkit_version:
+        print(
+            "Driver version %s.%s needs patching due to strings_udf"
+            % driver_version
+        )
+        if _numba_version_ok:
+            logger.debug("Patching Numba Linker")
+            Linker.new = new_patched_linker
+        else:
+            logger.debug("Cannot patch Numba Linker - unsupported version")
